pyproject.toml
[project]
name = "satellite-yolo-blackwell"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pyyaml>=6.0",
    "wandb>=0.16.0",
    "ultralytics>=8.1.0",
    "dvc[gdrive]>=3.0",
    "albumentations>=1.3.1",
    "opencv-python-headless>=4.8.0",
    "torch",
    "torchvision",
    "torchaudio",
    "datasets==2.19.1",
]

[[tool.uv.index]]
name = "pytorch-test"
url = "https://download.pytorch.org/whl/test/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-test" }]
torchvision = [{ index = "pytorch-test" }]
torchaudio = [{ index = "pytorch-test" }]
